backend/pyproject.toml
[project]
name = "asos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "bcrypt>=4.2.0",
    "fastapi>=0.115.2",
    "passlib>=1.7.4",
    "pydantic-settings>=2.6.0",
    "pydantic[email]>=2.9.2",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.12",
    "sqlmodel>=0.0.22",
]

[tool.uv]
dev-dependencies = [
    "fastapi-cli>=0.0.5",
]
